requirements.txt
mkdocs>=1.5
mkdocs-material>=9.5
pymdown-extensions>=10.8
mkdocs-awesome-pages-plugin>=2.10
mkdocs-macros-plugin>=1.5
Pillow>=10.0
